+++ module-external/src/main/java/com/yapp/sharefood/config/aws/LocalEmbeddedS3Config.java
@@ -5,23 +5,19 @@ import com.amazonaws.auth.AnonymousAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.yapp.sharefood.common.utils.ValidPortUtils;
 import io.findify.s3mock.S3Mock;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
-import org.springframework.util.StringUtils;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 
 @Slf4j
 @Profile({"local", "test"})
@@ -45,7 +41,7 @@ public class LocalEmbeddedS3Config {
 
     @Bean
     public S3Mock s3Mock() throws IOException {
-        port = AwsMockUtils.isRunningPort(port) ? AwsMockUtils.findAvailableRandomPort() : port;
+        port = ValidPortUtils.isRunningPort(port) ? ValidPortUtils.findAvailableRandomPort() : port;
         return new S3Mock.Builder()
                 .withPort(port)
                 .withInMemoryBackend()
@@ -84,54 +80,5 @@ public class LocalEmbeddedS3Config {
                 .port(port)
                 .build()
                 .toUriString();
-    }
-
-    @NoArgsConstructor(access = AccessLevel.PRIVATE)
-    public static final class AwsMockUtils {
-
-        private static final String OS = System.getProperty("os.name").toLowerCase();
-
-        public static boolean isRunningPort(int port) throws IOException {
-            return isRunning(executeGrepProcessCommand(port));
-        }
-
-        public static int findAvailableRandomPort() throws IOException {
-            for (int port = 10000; port <= 65535; port++) {
-                Process process = executeGrepProcessCommand(port);
-                if (!isRunning(process)) {
-                    return port;
-                }
-            }
-            throw new IllegalArgumentException("사용가능한 포트를 찾을 수 없습니다. (10000 ~ 65535)");
-        }
-
-        private static Process executeGrepProcessCommand(int port) throws IOException {
-            // 윈도우일 경우
-            if (isWindows()) {
-                String command = String.format("netstat -nao | find \"LISTEN\" | find \"%d\"", port);
-                String[] shell = {"cmd.exe", "/y", "/c", command};
-                return Runtime.getRuntime().exec(shell);
-            }
-            String command = String.format("netstat -nat | grep LISTEN|grep %d", port);
-            String[] shell = {"/bin/sh", "-c", command};
-            return Runtime.getRuntime().exec(shell);
-        }
-
-        private static boolean isWindows() {
-            return OS.contains("win");
-        }
-
-        private static boolean isRunning(Process process) {
-            String line;
-            StringBuilder pidInfo = new StringBuilder();
-            try (BufferedReader input = new BufferedReader(new InputStreamReader(process.getInputStream()))) {
-                while ((line = input.readLine()) != null) {
-                    pidInfo.append(line);
-                }
-            } catch (Exception e) {
-                throw new IllegalArgumentException("사용가능한 포트를 찾는 중 에러가 발생하였습니다.");
-            }
-            return StringUtils.hasLength(pidInfo.toString());
-        }
     }
 }
